Move Contour layer from experimental-layers to layers
