+++ modules/experimental-layers/src/index.js
@@ -17,6 +17,4 @@ export {default as GPUGridLayer} from './gpu-grid-layer/gpu-grid-layer';
 
 export {default as BitmapLayer} from './bitmap-layer/bitmap-layer';
 
-export {default as ContourLayer} from './contour-layer/contour-layer';
-
 export {default as TripsLayer} from './trips-layer/trips-layer';
