+++ modules/layers/src/contour-layer/marching-squares.js
