+++ modules/layers/src/index.js
@@ -39,6 +39,8 @@ export {default as GeoJsonLayer} from './geojson-layer/geojson-layer';
 
 export {default as TextLayer} from './text-layer/text-layer';
 
+export {default as ContourLayer} from './contour-layer/contour-layer';
+
 // Experimental layer exports
 export {default as _SolidPolygonLayer} from './solid-polygon-layer/solid-polygon-layer';
 export {default as _MultiIconLayer} from './text-layer/multi-icon-layer/multi-icon-layer';
